Moving create function to storageclassclaim module
