+++ ocs_ci/helpers/helpers.py
@@ -34,7 +34,7 @@ from ocs_ci.ocs.exceptions import (
     UnexpectedBehaviour,
 )
 from ocs_ci.ocs.ocp import OCP
-from ocs_ci.ocs.resources import pod, pvc, storageclassclaim
+from ocs_ci.ocs.resources import pod, pvc
 from ocs_ci.ocs.resources.ocs import OCS
 from ocs_ci.utility import templating
 from ocs_ci.utility.retry import retry
@@ -3909,46 +3909,3 @@ def get_noobaa_db_used_space():
         f"noobaa_db used space is {df_out[-4]} which is {df_out[-2]} of the total PVC size"
     )
     return df_out[-4]
-
-
-def create_storageclassclaim(
-    interface_type,
-    storage_class_claim_name=None,
-    namespace=None,
-):
-    """
-    Create a storageclassclaim
-
-    Args:
-        interface_type (str): The type of the interface
-            (e.g. CephBlockPool, CephFileSystem)
-        storage_class_claim_name (str): The name of storageclassclaim to create
-        namespace(str): The namespace in which the storageclassclaim should be created
-
-    Returns:
-        OCS: An OCS instance for the storageclassclaim
-    """
-    template_yaml = os.path.join(
-        constants.TEMPLATE_DIR, "storageclassclaim", "storageclassclaim.yaml"
-    )
-    sc_claim_data = templating.load_yaml(template_yaml)
-
-    if interface_type == constants.CEPHBLOCKPOOL:
-        type = "blockpool"
-    elif interface_type == constants.CEPHFILESYSTEM:
-        type = "sharedfilesystem"
-
-    sc_claim_data["spec"]["type"] = type
-    sc_claim_data["metadata"]["name"] = (
-        storage_class_claim_name
-        if storage_class_claim_name
-        else create_unique_resource_name(
-            f"test-{interface_type}", constants.STORAGECLASSCLAIM.lower()
-        )
-    )
-    if namespace:
-        sc_claim_data["metadata"]["namespace"] = namespace
-
-    sc_claim_obj = storageclassclaim.StorageClassClaim(**sc_claim_data)
-    created_sc_claim = sc_claim_obj.create(do_reload=True)
-    return created_sc_claim
